requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.4
boto3==1.7.62
botocore==1.10.62
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
Django==2.0.7
django-environ==0.4.5
django-politico-oembed-service==0.0.4
django-politico-token-service==0.0.3
django-s3-image-service==0.0.3
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
Fabric3==1.14.post1
flake8==3.5.0
idna==2.7
Jinja2==2.10
jmespath==0.9.3
kombu==4.2.1
MarkupSafe==1.0
mccabe==0.6.1
paramiko==2.4.1
Pillow==5.2.0
-e git+https://github.com/The-Politico/politico-services-api.git@55570b7c56c6bdd7b8f0ead8e848c4e884662c26#egg=politico_services_api
psycopg2==2.7.5
pyasn1==0.4.3
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyNaCl==1.2.1
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
vine==1.1.4
whitenoise==3.3.1
